+++ packages/core/types/utils/async.d.ts
@@ -1,2 +1,0 @@
-import { InternalRoute } from "../types/route";
-export default function hasAsyncRoute(routes: Array<InternalRoute>): boolean;