+++ packages/react/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next
 
+* Remove `<Link active>`.
 * Update `<Active>` to use a render-invoked `children` prop. Removes the `merge` prop.
 
 ## 1.0.0-beta.21